+++ codeforlife/tests/__init__.py
@@ -1,2 +1,9 @@
-from .api import APITestCase, APIClient
-from .cron import CronTestCase, CronTestClient
+"""
+© Ocado Group
+Created on 19/01/2024 at 17:17:23(+00:00).
+
+Custom test cases.
+"""
+
+from .cron import CronTestCase
+from .model_view_set import ModelViewSetTestCase
