+++ codeforlife/tests/cron.py
@@ -1,11 +1,20 @@
-from .api import APIClient, APITestCase
+"""
+© Ocado Group
+Created on 20/01/2024 at 09:52:43(+00:00).
+"""
+
+from rest_framework.test import APIClient, APITestCase
 
 
-class CronTestClient(APIClient):
+class CronClient(APIClient):
+    """Base client for all CRON jobs."""
+
     def __init__(self, *args, **kwargs):
         super().__init__(*args, **kwargs, HTTP_X_APPENGINE_CRON="true")
 
 
 class CronTestCase(APITestCase):
-    client: CronTestClient
-    client_class = CronTestClient
+    """Base test case for all CRON jobs."""
+
+    client: CronClient
+    client_class = CronClient
